hwc: Samsung hwc workaround: Fix blurryness in portrait
Apply the work around to landscape orientation only.

In portrait the display frame rect remains in landscape coordinates.
The source crop rect is in portrait. So the padding is applied to the side
and bottom causing the surface texture to become stretched and blurry.

It appears that the work around isn't necessary in portrait anyways.

(cherry picked from commit 22889d8cb433a373d95b8b02228fc73fdae565bc)