The real camera fix?!
I noticed that with the stock libjpeg.so, the camera takes good pictures,
but the device doesn't display any jpegs anymore. To see the pictures you need
to copy them to a different device. So, there's something in that libjpeg that
makes the camera work but makes the device not work. I really needed both
stock libjpeg.so, for the camera, and the one compiled from source to display
jpeg files. So, I went on a limb, renamed stock libjpeg to libtjpg.so and
used sed to change reference to linbjpeg.so to libtjpg.so in the camera hal:
sed -i s/libjpeg/libtjpg/ camera.tegra.so
sed -i s/libjpeg/libtjpg/ libtjpg.so
and that seems to work.

Change-Id: I88bd95771519a2e8d2fe102461a904f9c938f86e

Conflicts:
	BoardConfig.mk