Assorted fixes: new kernel, no more kernel modules, no more conservative cpufreq-gov
The new kernel has all device drivers built-in so that module loading during init
is no longer necessary, which silences the last of the selinux warnings.
Additionally, this, together with the earlier removal of NvCPLSvc module,
and rewriting the board architecture from tegra3 to tegra, removes the necessity
of mounting the system partition rw and then remounting it, and removes the
need to write symlinks in /system/lib/hw (*tegra* -> *tegra3*). Finally,
the new kernel no longer switches to the conservative CPU frequency governor,
which makes it easier to reuse the grouper power hal, which I hope will reduce
the risk of wakelock issues

Change-Id: Ib798dcad9c0217a5861311f17b0c1403b7952f39