Reset kernel, init, proprietary files, liblights and powerhal
This commit reverts to the kitkat kernel, removes the open source
lights hal and open source power hal, and largely reverts to stock
nvidia libs (no grouper libs). This in order to be as kitkat like
as possible, and not to brick any more devices. I also removed the
camera wrapper, as I think it should not be necessary.

Still, some potential risks remain, such as the removal of NvCPLSvc.

Change-Id: I662c5be7e339bcc59320d5d7a77d00623a348678